+++ src/array/ft_array_set.c
@@ -6,13 +6,13 @@
 /*   By: abrabant <abrabant@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/03/22 12:31:10 by abrabant          #+#    #+#             */
-/*   Updated: 2021/03/22 12:32:17 by abrabant         ###   ########.fr       */
+/*   Updated: 2021/03/27 15:06:57 by abrabant         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft/internal/array_int.h"
 
-void	ft_array_set(t_array_int *a, void *el, int i)
+void	ft_array_set(t_array_int *a, void *el, size_t i)
 {
 	a->array[i] = el;
 }
